requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
SQLAlchemy>=2.0.16
python-dotenv==1.0.1
pytest==8.0.2
bcrypt==4.1.2
marshmallow==3.20.2
python-dateutil==2.8.2
alembic==1.13.1
Werkzeug==3.0.1
flask-restx==1.3.0 